+++ pdtable/io/excel.py
@@ -8,7 +8,6 @@ requiring them (read_excel() or write_excel()) are called for the first time.
 
 """
 import os
-import io
 from os import PathLike
 from typing import Union, Callable, Iterable, BinaryIO
 
@@ -53,15 +52,12 @@ def read_excel(
             "Please install openpyxl for Excel I/O support."
         ) from err
 
-    if not isinstance(source, (str, PathLike)):
-        assert isinstance(source, io.BufferedIOBase)
-
     yield from parse_blocks(read_cell_rows(source), **kwargs)
 
 
 def write_excel(
     tables: Union[Table, Iterable[Table], TableBundle],
-    path: Union[str, os.PathLike],
+    to: Union[str, os.PathLike, BinaryIO],
     na_rep: str = "-",
 ):
     """Writes one or more tables to an Excel workbook.
@@ -75,8 +71,11 @@ def write_excel(
     Args:
         tables:
             Table(s) to write. Can be a single Table or an iterable of Tables.
-        path:
-            File path to which to write.
+        to:
+            File path or binary stream to which to write.
+            If a file path, then this file gets created/overwritten and then closed after writing.
+            If a stream, then it is left open after writing; the caller is responsible for managing
+            the stream.
         na_rep:
             Optional; String representation of missing values (NaN, None, NaT).
             If overriding the default '-', it is recommended to use another value compliant with
@@ -92,4 +91,4 @@ def write_excel(
             "Please install openpyxl for Excel I/O support."
         ) from err
 
-    write_excel_func(tables, path, na_rep)
+    write_excel_func(tables, to, na_rep)